refactor

+++ app/src/main/java/com/eshc/travelplatform/data/plan/Spot.kt
@@ -1,3 +1,3 @@
 package com.eshc.travelplatform.data.plan
 
-data class Spot(val name: String, val x : Float, val y : Float)
+data class Spot(val name: String, val x : Double, val y : Double)
